moved DeviceSharedDofCommunicator inside GroupCommunicator.
