+++ fem/pfespace.cpp
@@ -106,7 +106,6 @@ void ParFiniteElementSpace::ParInit(ParMesh *pm)
    nonconf_P = false;
    Rconf = nullptr;
    R = nullptr;
-   DofComm.reset();
    num_face_nbr_dofs = -1;
 
    if (NURBSext && !pNURBSext())
@@ -1143,12 +1142,7 @@ ParFiniteElementSpace::GetDeviceSharedDofCommunicator() const
    MFEM_VERIFY(UseDeviceSharedDofComm(),
                "Shared-dof communicator requires a conforming space with "
                "1-1 shared dof correspondence.");
-   if (!DofComm)
-   {
-      Dof_TrueDof_Matrix(); // ensure R is built
-      DofComm = std::make_unique<DeviceSharedDofCommunicator>(*this);
-   }
-   return DofComm.get();
+   return gcomm->GetDeviceSharedDofCommunicator();
 }
 
 void ParFiniteElementSpace::Synchronize(Array<int> &ldof_marker) const
@@ -1830,7 +1824,6 @@ const FiniteElement *ParFiniteElementSpace::GetFaceNbrFaceFE(int i) const
 
 void ParFiniteElementSpace::Lose_Dof_TrueDof_Matrix()
 {
-   DofComm.reset();
    P -> StealData();
 #if MFEM_HYPRE_VERSION <= 22200
    hypre_ParCSRMatrix *csrP = (hypre_ParCSRMatrix*)(*P);
@@ -1889,6 +1882,7 @@ void ParFiniteElementSpace::ConstructTrueDofs()
       }
    }
    gcomm->SetLTDofTable(ldof_ltdof);
+   gcomm->SetDeviceSharedDofSupport(Conforming() && !SharedNDTriangleDofs());
 
    // have the group masters broadcast their ltdofs to the rest of the group
    gcomm->Bcast(ldof_ltdof);
@@ -1936,6 +1930,7 @@ void ParFiniteElementSpace::ConstructTrueNURBSDofs()
       }
    }
    gcomm->SetLTDofTable(ldof_ltdof);
+   gcomm->SetDeviceSharedDofSupport(Conforming() && !SharedNDTriangleDofs());
 
    // have the group masters broadcast their ltdofs to the rest of the group
    gcomm->Bcast(ldof_ltdof);
@@ -4807,7 +4802,6 @@ void ParFiniteElementSpace::Destroy()
    delete Pconf; Pconf = NULL;
    delete Rconf; Rconf = NULL;
    delete R; R = NULL;
-   DofComm.reset();
 
    delete gcomm; gcomm = NULL;
 
@@ -5301,261 +5295,10 @@ void ConformingProlongationOperator::MultTranspose(
    }
 }
 
-namespace internal
-{
-
-class DeviceNeighborDofComm
-{
-public:
-   DeviceNeighborDofComm(const GroupCommunicator &gc_, const SparseMatrix &R)
-      : gc(gc_),
-        mpi_gpu_aware(Device::GetGPUAwareMPI())
-   {
-      MFEM_ASSERT(R.Finalized(), "");
-      const int tdofs = R.Height();
-      MFEM_ASSERT(tdofs == R.HostReadI()[tdofs], "");
-      ltdof_ldof.SetSize(tdofs);
-      ltdof_ldof.CopyFrom(R.HostReadJ());
-      {
-         Table nbr_ltdof;
-         gc.GetNeighborLTDofTable(nbr_ltdof);
-         const int nb_connections = nbr_ltdof.Size_of_connections();
-         shr_ltdof.SetSize(nb_connections);
-         if (nb_connections > 0) { shr_ltdof.CopyFrom(nbr_ltdof.GetJ()); }
-         shr_buf_offsets = nbr_ltdof.GetIMemory();
-         {
-            Array<int> shared_ltdof(nbr_ltdof.GetJ(), nb_connections);
-            Array<int> unique_ltdof(shared_ltdof);
-            unique_ltdof.Sort();
-            unique_ltdof.Unique();
-            for (int i = 0; i < shared_ltdof.Size(); i++)
-            {
-               shared_ltdof[i] = unique_ltdof.FindSorted(shared_ltdof[i]);
-               MFEM_ASSERT(shared_ltdof[i] != -1, "internal error");
-            }
-            Table unique_shr;
-            Transpose(shared_ltdof, unique_shr, unique_ltdof.Size());
-            unq_ltdof = unique_ltdof;
-            unq_shr_i.GetMemory() = unique_shr.GetIMemory();
-            unq_shr_i.SetSize(unique_shr.Size()+1);
-            unq_shr_j.GetMemory() = unique_shr.GetJMemory();
-            unq_shr_j.SetSize(unique_shr.Size_of_connections());
-            unique_shr.LoseData();
-         }
-         nbr_ltdof.GetJMemory().Delete();
-         nbr_ltdof.LoseData();
-      }
-      {
-         Table nbr_ldof;
-         gc.GetNeighborLDofTable(nbr_ldof);
-         const int nb_connections = nbr_ldof.Size_of_connections();
-         ext_ldof.SetSize(nb_connections);
-         if (nb_connections > 0) { ext_ldof.CopyFrom(nbr_ldof.GetJ()); }
-         ext_ldof.GetMemory().UseDevice(true);
-         ext_buf_offsets = nbr_ldof.GetIMemory();
-         nbr_ldof.GetJMemory().Delete();
-         nbr_ldof.LoseData();
-      }
-      const GroupTopology &gtopo = gc.GetGroupTopology();
-      int req_counter = 0;
-      for (int nbr = 1; nbr < gtopo.GetNumNeighbors(); nbr++)
-      {
-         const int send_offset = shr_buf_offsets[nbr];
-         const int send_size = shr_buf_offsets[nbr+1] - send_offset;
-         if (send_size > 0) { req_counter++; }
-
-         const int recv_offset = ext_buf_offsets[nbr];
-         const int recv_size = ext_buf_offsets[nbr+1] - recv_offset;
-         if (recv_size > 0) { req_counter++; }
-      }
-      requests.resize(req_counter);
-   }
-
-   ~DeviceNeighborDofComm()
-   {
-      ext_buf_offsets.Delete();
-      shr_buf_offsets.Delete();
-   }
-
-   bool MpiGpuAware() const { return mpi_gpu_aware; }
-
-   const Array<int> &SharedLTDof() const { return shr_ltdof; }
-   const Array<int> &ExternalLDof() const { return ext_ldof; }
-   const Array<int> &LocalTDofToLDof() const { return ltdof_ldof; }
-   const Array<int> &UniqueLTDof() const { return unq_ltdof; }
-   const Array<int> &UniqueSharedOffsets() const { return unq_shr_i; }
-   const Array<int> &UniqueSharedIndices() const { return unq_shr_j; }
-
-   template <typename T, typename SendBuffer, typename RecvBuffer>
-   int ExchangeSharedToExternal(const SendBuffer &shr_buf,
-                                RecvBuffer &ext_buf,
-                                int tag) const
-   {
-      return Exchange<T>(shr_buf, shr_buf_offsets, ext_buf, ext_buf_offsets,
-                         tag);
-   }
-
-   template <typename T, typename SendBuffer, typename RecvBuffer>
-   int ExchangeExternalToShared(const SendBuffer &ext_buf,
-                                RecvBuffer &shr_buf,
-                                int tag) const
-   {
-      return Exchange<T>(ext_buf, ext_buf_offsets, shr_buf, shr_buf_offsets,
-                         tag);
-   }
-
-   void WaitAll(int req_counter) const
-   {
-      MPI_Waitall(req_counter, requests.data(), MPI_STATUSES_IGNORE);
-   }
-
-private:
-   template <typename T, typename SendBuffer, typename RecvBuffer>
-   int Exchange(const SendBuffer &send_buf, const Memory<int> &send_offsets,
-                RecvBuffer &recv_buf, const Memory<int> &recv_offsets,
-                int tag) const
-   {
-      const GroupTopology &gtopo = gc.GetGroupTopology();
-      int req_counter = 0;
-      for (int nbr = 1; nbr < gtopo.GetNumNeighbors(); nbr++)
-      {
-         const int send_offset = send_offsets[nbr];
-         const int send_size = send_offsets[nbr+1] - send_offset;
-         if (send_size > 0)
-         {
-            auto send_ptr = mpi_gpu_aware ? send_buf.Read() : send_buf.HostRead();
-            MPI_Isend(send_ptr + send_offset, send_size, MPITypeMap<T>::mpi_type,
-                      gtopo.GetNeighborRank(nbr), tag, gtopo.GetComm(),
-                      &requests[req_counter++]);
-         }
-         const int recv_offset = recv_offsets[nbr];
-         const int recv_size = recv_offsets[nbr+1] - recv_offset;
-         if (recv_size > 0)
-         {
-            auto recv_ptr = mpi_gpu_aware ? recv_buf.Write() : recv_buf.HostWrite();
-            MPI_Irecv(recv_ptr + recv_offset, recv_size, MPITypeMap<T>::mpi_type,
-                      gtopo.GetNeighborRank(nbr), tag, gtopo.GetComm(),
-                      &requests[req_counter++]);
-         }
-      }
-      return req_counter;
-   }
-
-   const GroupCommunicator &gc;
-   bool mpi_gpu_aware;
-   Array<int> shr_ltdof, ext_ldof;
-   Memory<int> shr_buf_offsets, ext_buf_offsets;
-   Array<int> ltdof_ldof, unq_ltdof;
-   Array<int> unq_shr_i, unq_shr_j;
-   mutable std::vector<MPI_Request> requests;
-};
-
-template <typename InBuffer, typename OutBuffer>
-void DeviceNeighborDofExtract(const Array<int> &indices,
-                              const InBuffer &xin,
-                              OutBuffer &xout)
-{
-   MFEM_ASSERT(indices.Size() == xout.Size(), "incompatible sizes!");
-   auto y = xout.Write();
-   const auto x = xin.Read();
-   const auto I = indices.Read();
-   mfem::forall(indices.Size(), [=] MFEM_HOST_DEVICE (int i)
-   {
-      y[i] = x[I[i]];
-   }); // indices can be repeated
-}
-
-template <typename InBuffer, typename OutBuffer>
-void DeviceNeighborDofSet(const Array<int> &indices,
-                          const InBuffer &xin,
-                          OutBuffer &xout)
-{
-   MFEM_ASSERT(indices.Size() == xin.Size(), "incompatible sizes!");
-   auto y = xout.ReadWrite();
-   const auto x = xin.Read();
-   const auto I = indices.Read();
-   mfem::forall(indices.Size(), [=] MFEM_HOST_DEVICE (int i)
-   {
-      y[I[i]] = x[i];
-   });
-}
-
-template <typename SrcBuffer, typename DstBuffer>
-void DeviceNeighborDofAdd(const Array<int> &unique_dst_indices,
-                          const Array<int> &unique_to_src_offsets,
-                          const Array<int> &unique_to_src_indices,
-                          const SrcBuffer &src,
-                          DstBuffer &dst)
-{
-   auto y = dst.ReadWrite();
-   const auto x = src.Read();
-   const auto DST_I = unique_dst_indices.Read();
-   const auto SRC_O = unique_to_src_offsets.Read();
-   const auto SRC_I = unique_to_src_indices.Read();
-   mfem::forall(unique_dst_indices.Size(), [=] MFEM_HOST_DEVICE (int i)
-   {
-      const int dst_idx = DST_I[i];
-      real_t sum = y[dst_idx];
-      const int end = SRC_O[i+1];
-      for (int j = SRC_O[i]; j != end; ++j) { sum += x[SRC_I[j]]; }
-      y[dst_idx] = sum;
-   });
-}
-
-template <typename T>
-void DeviceSharedDofApplyReduction(const Array<int> &unique_dst_indices,
-                                   const Array<int> &unique_to_src_offsets,
-                                   const Array<int> &unique_to_src_indices,
-                                   const Array<T> &src,
-                                   Array<T> &dst,
-                                   DeviceSharedDofCommunicator::Op op)
-{
-   auto y = dst.ReadWrite();
-   const auto x = src.Read();
-   const auto DST_I = unique_dst_indices.Read();
-   const auto SRC_O = unique_to_src_offsets.Read();
-   const auto SRC_I = unique_to_src_indices.Read();
-   mfem::forall(unique_dst_indices.Size(), [=] MFEM_HOST_DEVICE (int i)
-   {
-      const int dst_idx = DST_I[i];
-      T val = y[dst_idx];
-      const int end = SRC_O[i+1];
-      switch (op)
-      {
-         case DeviceSharedDofCommunicator::Op::Sum:
-         {
-            for (int j = SRC_O[i]; j != end; ++j) { val += x[SRC_I[j]]; }
-            break;
-         }
-         case DeviceSharedDofCommunicator::Op::Min:
-         {
-            for (int j = SRC_O[i]; j != end; ++j)
-            {
-               const T xj = x[SRC_I[j]];
-               val = (xj < val) ? xj : val;
-            }
-            break;
-         }
-         case DeviceSharedDofCommunicator::Op::Max:
-         {
-            for (int j = SRC_O[i]; j != end; ++j)
-            {
-               const T xj = x[SRC_I[j]];
-               val = (xj > val) ? xj : val;
-            }
-            break;
-         }
-      }
-      y[dst_idx] = val;
-   });
-}
-
-} // namespace internal
-
 DeviceConformingProlongationOperator::DeviceConformingProlongationOperator(
-   const GroupCommunicator &gc_, const SparseMatrix *R, bool local_)
-   : ConformingProlongationOperator(R->Width(), gc_, local_),
-     nbr_comm(new internal::DeviceNeighborDofComm(gc_, *R))
+   int lsize, const GroupCommunicator &gc_, bool local_)
+   : ConformingProlongationOperator(lsize, gc_, local_),
+     nbr_comm(new internal::DeviceNeighborDofComm(gc_))
 {
    shr_buf.SetSize(nbr_comm->SharedLTDof().Size());
    shr_buf.UseDevice(true);
@@ -5565,12 +5308,11 @@ DeviceConformingProlongationOperator::DeviceConformingProlongationOperator(
 
 DeviceConformingProlongationOperator::DeviceConformingProlongationOperator(
    const ParFiniteElementSpace &pfes, bool local_)
-   : DeviceConformingProlongationOperator(pfes.GroupComm(),
-                                          pfes.GetRestrictionMatrix(),
+   : DeviceConformingProlongationOperator(pfes.GetVSize(),
+                                          pfes.GroupComm(),
                                           local_)
 {
    MFEM_ASSERT(pfes.Conforming(), "internal error");
-   MFEM_ASSERT(pfes.GetRestrictionMatrix()->Height() == pfes.GetTrueVSize(), "");
 }
 
 void DeviceConformingProlongationOperator::BcastBeginCopy(
@@ -5679,235 +5421,6 @@ void DeviceConformingProlongationOperator::MultTranspose(const Vector &x,
       ReduceEndAssemble(y); // assemble from 'shr_buf'
    }
 }
-
-template <typename T>
-void DeviceSharedDofCommunicator::ReduceBeginCopy(const Array<T> &x_ldof,
-                                                  Array<T> &ext_buf_t) const
-{
-   const auto &ext_ldof = nbr_comm->ExternalLDof();
-   if (ext_ldof.Size() == 0) { return; }
-   internal::DeviceNeighborDofExtract(ext_ldof, x_ldof, ext_buf_t);
-   if (nbr_comm->MpiGpuAware()) { MFEM_STREAM_SYNC; }
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::ReduceLocalCopy(const Array<T> &x_ldof,
-                                                  Array<T> &x_tdof) const
-{
-   const auto &ltdof_ldof = nbr_comm->LocalTDofToLDof();
-   if (ltdof_ldof.Size() == 0) { return; }
-   internal::DeviceNeighborDofExtract(ltdof_ldof, x_ldof, x_tdof);
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::ReduceEndAssemble(const Array<T> &shr_buf_t,
-                                                    Array<T> &x_tdof,
-                                                    Op op) const
-{
-   const auto &unq_ltdof = nbr_comm->UniqueLTDof();
-   if (unq_ltdof.Size() == 0) { return; }
-   internal::DeviceSharedDofApplyReduction(unq_ltdof,
-                                           nbr_comm->UniqueSharedOffsets(),
-                                           nbr_comm->UniqueSharedIndices(),
-                                           shr_buf_t, x_tdof, op);
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::BcastBeginCopy(const Array<T> &x_tdof,
-                                                 Array<T> &shr_buf_t) const
-{
-   const auto &shr_ltdof = nbr_comm->SharedLTDof();
-   if (shr_ltdof.Size() == 0) { return; }
-   internal::DeviceNeighborDofExtract(shr_ltdof, x_tdof, shr_buf_t);
-   if (nbr_comm->MpiGpuAware()) { MFEM_STREAM_SYNC; }
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::BcastLocalCopy(const Array<T> &x_tdof,
-                                                 Array<T> &x_ldof) const
-{
-   const auto &ltdof_ldof = nbr_comm->LocalTDofToLDof();
-   if (ltdof_ldof.Size() == 0) { return; }
-   internal::DeviceNeighborDofSet(ltdof_ldof, x_tdof, x_ldof);
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::BcastEndCopy(const Array<T> &ext_buf_t,
-                                               Array<T> &x_ldof) const
-{
-   const auto &ext_ldof = nbr_comm->ExternalLDof();
-   if (ext_ldof.Size() == 0) { return; }
-   internal::DeviceNeighborDofSet(ext_ldof, ext_buf_t, x_ldof);
-}
-
-DeviceSharedDofCommunicator::DeviceSharedDofCommunicator(
-   const ParFiniteElementSpace &pfes)
-   : nbr_comm(new internal::DeviceNeighborDofComm(pfes.GroupComm(),
-                                                  *pfes.GetRestrictionMatrix()))
-{
-   const SparseMatrix *R = pfes.GetRestrictionMatrix();
-   // Temporary storage for true-dof data during reduce+broadcast sequences.
-   true_buf.SetSize(R->Height());
-   true_buf.GetMemory().UseDevice(true);
-   shr_buf.SetSize(nbr_comm->SharedLTDof().Size());
-   shr_buf.GetMemory().UseDevice(true);
-   ext_buf.SetSize(nbr_comm->ExternalLDof().Size());
-   ext_buf.GetMemory().UseDevice(true);
-   MFEM_ASSERT(pfes.Conforming(), "internal error");
-   MFEM_ASSERT(pfes.GetRestrictionMatrix()->Height() == pfes.GetTrueVSize(), "");
-}
-
-DeviceSharedDofCommunicator::~DeviceSharedDofCommunicator() = default;
-
-template <typename T>
-void DeviceSharedDofCommunicator::Reduce(const Array<T> &x_ldof,
-                                         Array<T> &x_tdof,
-                                         Op op) const
-{
-   MFEM_ASSERT(x_tdof.Size() == nbr_comm->LocalTDofToLDof().Size(),
-               "incompatible sizes!");
-   Array<T> ext_buf_t(nbr_comm->ExternalLDof().Size());
-   Array<T> shr_buf_t(nbr_comm->SharedLTDof().Size());
-   ext_buf_t.GetMemory().UseDevice(true);
-   shr_buf_t.GetMemory().UseDevice(true);
-   ReduceBeginCopy(x_ldof, ext_buf_t);
-   const int req_counter =
-      nbr_comm->ExchangeExternalToShared<T>(ext_buf_t, shr_buf_t, 41827);
-   ReduceLocalCopy(x_ldof, x_tdof);
-   nbr_comm->WaitAll(req_counter);
-   ReduceEndAssemble(shr_buf_t, x_tdof, op);
-}
-
-template <>
-void DeviceSharedDofCommunicator::Reduce<real_t>(const Array<real_t> &x_ldof,
-                                                 Array<real_t> &x_tdof,
-                                                 Op op) const
-{
-   MFEM_ASSERT(x_tdof.Size() == nbr_comm->LocalTDofToLDof().Size(),
-               "incompatible sizes!");
-   ReduceBeginCopy(x_ldof, ext_buf);
-   const int req_counter =
-      nbr_comm->ExchangeExternalToShared<real_t>(ext_buf, shr_buf, 41825);
-   ReduceLocalCopy(x_ldof, x_tdof);
-   nbr_comm->WaitAll(req_counter);
-   ReduceEndAssemble(shr_buf, x_tdof, op);
-}
-
-void DeviceSharedDofCommunicator::Reduce(const Vector &x_ldof,
-                                         Vector &x_tdof,
-                                         Op op) const
-{
-   Array<real_t> x_ldof_view, x_tdof_view;
-   x_ldof_view.MakeRef(const_cast<Memory<real_t>&>(x_ldof.GetMemory()), 0,
-                       x_ldof.Size());
-   x_tdof_view.MakeRef(const_cast<Memory<real_t>&>(x_tdof.GetMemory()), 0,
-                       x_tdof.Size());
-   Reduce(x_ldof_view, x_tdof_view, op);
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::Bcast(const Array<T> &x_tdof,
-                                        Array<T> &x_ldof) const
-{
-   MFEM_ASSERT(x_tdof.Size() == nbr_comm->LocalTDofToLDof().Size(),
-               "incompatible sizes!");
-   Array<T> ext_buf_t(nbr_comm->ExternalLDof().Size());
-   Array<T> shr_buf_t(nbr_comm->SharedLTDof().Size());
-   ext_buf_t.GetMemory().UseDevice(true);
-   shr_buf_t.GetMemory().UseDevice(true);
-   x_ldof.Write();
-   BcastBeginCopy(x_tdof, shr_buf_t);
-   const int req_counter =
-      nbr_comm->ExchangeSharedToExternal<T>(shr_buf_t, ext_buf_t, 41826);
-   BcastLocalCopy(x_tdof, x_ldof);
-   nbr_comm->WaitAll(req_counter);
-   BcastEndCopy(ext_buf_t, x_ldof);
-}
-
-template <>
-void DeviceSharedDofCommunicator::Bcast<real_t>(const Array<real_t> &x_tdof,
-                                                Array<real_t> &x_ldof) const
-{
-   MFEM_ASSERT(x_tdof.Size() == nbr_comm->LocalTDofToLDof().Size(),
-               "incompatible sizes!");
-   x_ldof.Write();
-   BcastBeginCopy(x_tdof, shr_buf);
-   const int req_counter =
-      nbr_comm->ExchangeSharedToExternal<real_t>(shr_buf, ext_buf, 41824);
-   BcastLocalCopy(x_tdof, x_ldof);
-   nbr_comm->WaitAll(req_counter);
-   BcastEndCopy(ext_buf, x_ldof);
-}
-
-void DeviceSharedDofCommunicator::Bcast(const Vector &x_tdof,
-                                        Vector &x_ldof) const
-{
-   Array<real_t> x_tdof_view, x_ldof_view;
-   x_tdof_view.MakeRef(const_cast<Memory<real_t>&>(x_tdof.GetMemory()), 0,
-                       x_tdof.Size());
-   x_ldof_view.MakeRef(const_cast<Memory<real_t>&>(x_ldof.GetMemory()), 0,
-                       x_ldof.Size());
-   Bcast(x_tdof_view, x_ldof_view);
-}
-
-template <typename T>
-void DeviceSharedDofCommunicator::ReduceAndBcast(Array<T> &x_ldof, Op op) const
-{
-   Array<T> x_tdof(nbr_comm->LocalTDofToLDof().Size());
-   x_tdof.GetMemory().UseDevice(true);
-   Reduce(x_ldof, x_tdof, op);
-   Bcast(x_tdof, x_ldof);
-}
-
-template <>
-void DeviceSharedDofCommunicator::ReduceAndBcast<real_t>(Array<real_t> &x_ldof,
-                                                         Op op) const
-{
-   Reduce(x_ldof, true_buf, op);
-   Bcast(true_buf, x_ldof);
-}
-
-void DeviceSharedDofCommunicator::ReduceAndBcast(Vector &x_ldof, Op op) const
-{
-   Array<real_t> x_ldof_view;
-   x_ldof_view.MakeRef(const_cast<Memory<real_t>&>(x_ldof.GetMemory()), 0,
-                       x_ldof.Size());
-   ReduceAndBcast(x_ldof_view, op);
-}
-
-// @cond DOXYGEN_SKIP
-template void DeviceSharedDofCommunicator::ReduceBeginCopy<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::ReduceLocalCopy<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::ReduceEndAssemble<int>(
-   const Array<int> &, Array<int> &, Op) const;
-template void DeviceSharedDofCommunicator::BcastBeginCopy<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::BcastLocalCopy<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::BcastEndCopy<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::Reduce<int>(
-   const Array<int> &, Array<int> &, Op) const;
-template void DeviceSharedDofCommunicator::Bcast<int>(
-   const Array<int> &, Array<int> &) const;
-template void DeviceSharedDofCommunicator::ReduceAndBcast<int>(
-   Array<int> &, Op) const;
-
-template void DeviceSharedDofCommunicator::ReduceBeginCopy<real_t>(
-   const Array<real_t> &, Array<real_t> &) const;
-template void DeviceSharedDofCommunicator::ReduceLocalCopy<real_t>(
-   const Array<real_t> &, Array<real_t> &) const;
-template void DeviceSharedDofCommunicator::ReduceEndAssemble<real_t>(
-   const Array<real_t> &, Array<real_t> &, Op) const;
-template void DeviceSharedDofCommunicator::BcastBeginCopy<real_t>(
-   const Array<real_t> &, Array<real_t> &) const;
-template void DeviceSharedDofCommunicator::BcastLocalCopy<real_t>(
-   const Array<real_t> &, Array<real_t> &) const;
-template void DeviceSharedDofCommunicator::BcastEndCopy<real_t>(
-   const Array<real_t> &, Array<real_t> &) const;
-// @endcond
 
 } // namespace mfem
